+++ packages/create/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/create
 
+## 0.8.16
+
+### Patch Changes
+
+-   Updated dependencies [2fd82b1]
+    -   @sap-ux/adp-tooling@0.12.48
+    -   @sap-ux/preview-middleware@0.16.65
+
+## 0.8.15
+
+### Patch Changes
+
+-   Updated dependencies [09f91c3]
+    -   @sap-ux/preview-middleware@0.16.64
+
 ## 0.8.14
 
 ### Patch Changes